Refactor: Created a lib with the turing logic
